Implemented Joystick Control on Quay Crane
Changelog:
Implemented joystick crane control. It is now possible to control the crane using the joystick

Renamed boundry boxes, they were named incorrectly and thus not detected by the boundry part of the code

Added a boundry box for the forward part of the crane.

List of current issues:
Auditory feedback reacts to the boundry boxes, beeing despite there being nothing close by.

Boundry boxes are currently only reacting to the main crane, not the trolley, beacuse of the way they are being compared in the code.

The forward boundry box is currently coliding with the crane at start, causing any lateral movement to not function. Leaving it as is because i need to rebuild the boundry detection system anyway

Code is spitting out massive amounts of index out of bounds and nullrefrence exception errors. All of this is comming from the boat AI part of the program.

Cameras you implemented is currently not displaying when start is hit. Will investigate.