Bugfixes and Crane feedback implementation
Changelog:
Removed collision GameObjects forward, backward, left and right.

Added Min and Max values for X and Y to the crane, the X value working as the crane safe area and the Y value working as the trolley safe area

Implemented joystick force feedback support when moving container onto ship. The force is also progressive, so essentially one could let go of the stick and it will guide the container to its target on its own. There is a margin of error, its small enough to where i doubt anyone will care.

Removed your raycast collision function, aswell as the stop booleans, cause i was getting cancer from the latter and the former was no longer needed anyway.

Disabled the boat AI, it was just spitting out errors and not needed anyway.

Fixed the truck despawn issue.

Known Issues:
We have too many audio listeners, the log is blowing up with complaints about that to the point where working is almost impossible. Idk where they are so i cant fix it myself

Container is not properly detaching from trolley, when detached it lets go of the trolley but the trolley can still steer it. Im not sure where to even begin to fix that one.

Pressing C changes the camera, when let go it causes issues and entirely locks up and disables controls.

Containers are currently only colliding with containers, and nothing else. Is this by design?

Force feedback when picking up container off truck not implemented yet, unsure how to get the location of the container on that truck.